+++ app/src/main/java/jgame/gradle/CircusCharlie/FXPlayer.java
@@ -80,7 +80,7 @@ public enum FXPlayer {
       }
    }
  
-   static void init() {
+   public static void init() {
       values();
    }
 }
